requirements.txt
discord.py
yt-dlp
g4f
python-dotenv
flask